Parametrization interface.
