+++ cnqr/layers.py
@@ -1,6 +1,6 @@
-from cnqr._src.lipschitz import StiefelDense
 from cnqr._src.lipschitz import Normalized2ToInftyDense
 from cnqr._src.lipschitz import NormalizedInftyDense
-from cnqr._src.lipschitz import full_sort
+from cnqr._src.lipschitz import StiefelDense
+from cnqr._src.lipschitz import fullsort
 from cnqr._src.lipschitz import groupsort2
 from cnqr._src.convex import PICNN
