+++ cnqr/parametrizations.py
@@ -1,0 +1,3 @@
+from cnqr._src.parametrizations import BjorckParametrization
+from cnqr._src.parametrizations import Normalized2ToInftyParametrization
+from cnqr._src.parametrizations import NormalizedInftyParametrization
