+++ examples/ot/wasserstein_1.py
@@ -20,37 +20,35 @@ from jax.config import config
 from flax.training import train_state
 from flax import linen as nn
 from matplotlib import axes
+import numpy as onp
 
 import optax
 from sklearn.datasets import make_moons, make_circles
 
 from cnqr import balanced_KR
-from cnqr.layers import StiefelDense, full_sort, groupsort2
+from cnqr.layers import StiefelDense, Normalized2ToInftyDense, fullsort, groupsort2
 
 
-flags.DEFINE_float("learning_rate", 1e-3, "Learning rate.")
+flags.DEFINE_float("learning_rate", 1e-2, "Learning rate.")
 flags.DEFINE_float("noise", 0.05, "Noise in observations.")
-flags.DEFINE_integer("dataset_size", 5000, "Number of points in dataset.")
-flags.DEFINE_integer("num_epochs", 50, "Number of passes over the dataset.")
+flags.DEFINE_integer("dataset_size", 4096, "Number of points in dataset.")
+flags.DEFINE_integer("num_epochs", 200, "Number of passes over the dataset.")
 flags.DEFINE_enum("dataset", "two_moons", ['two_moons', 'two_circles'], "Dataset to train on.")
 flags.DEFINE_integer("batch_size", 256, "Number of examples in batch.")
 FLAGS = flags.FLAGS
 
 
 class LipschitzNN(nn.Module):
-  hidden_widths: Sequence[int] = (64, 64, 64, 64)
+  hidden_widths: Sequence[int] = (256, 128, 64)
 
   @nn.compact
   def __call__(self, inputs, train):
-    stiefel_dense = partial(
-      StiefelDense,
-      train=train)
 
     x = inputs
     for width in self.hidden_widths:
-      x = stiefel_dense(features=width)(x)
-      x = full_sort(x)
-    x = stiefel_dense(features=1)(x)
+      x = StiefelDense(features=width)(x, train=train)  
+      x = fullsort(x)
+    x = Normalized2ToInftyDense(features=1, use_bias=False)(x, train=train)
 
     return x
 
@@ -133,7 +131,7 @@ def train_epoch(state, train_ds, batch_size, rng):
     epoch_loss.append(loss)
 
   train_loss = -jnp.mean(jnp.array(epoch_loss))
-  return state, train_loss
+  return state, train_loss, (P.sum(), Q.sum())
 
 
 def get_datasets():
@@ -150,16 +148,42 @@ def get_datasets():
   return train_ds
 
 
-def compute_earth_mover_distance(points, labels, epsilon):
+def compute_earth_mover_distance_ott(points, labels, epsilon=1e-2):
+  """Compute the Earth Mover's Distance (EMD) between two distributions.
+
+  Based on the OTT library: https://ott-jax.readthedocs.io
+  """
   import ott
-  from ott.geometry import pointcloud
-  from ott.core import sinkhorn
-  from ott.tools import transport
+  from ott.geometry import costs, pointcloud
+  from ott.problems.linear import linear_problem
+  from ott.solvers.linear import sinkhorn
   print("Compute EMD...")
   p, q = points[labels == 1], points[labels == -1]
-  geom = pointcloud.PointCloud(p, q, epsilon=epsilon, power=1.0)
-  out = sinkhorn.sinkhorn(geom)
-  return out.reg_ot_cost
+  geom = pointcloud.PointCloud(p, q, epsilon=epsilon,
+                               cost_fn=costs.Euclidean)
+  ot_prob = linear_problem.LinearProblem(geom)
+  solver = sinkhorn.Sinkhorn()
+  out = solver(ot_prob)
+  emd = out.reg_ot_cost
+  print(f"Regulized EMD (ε={epsilon:.2f}): {float(emd):.2f}")
+  return emd
+
+
+def compute_earth_mover_distance_pot(points, labels):
+  """Compute the Earth Mover's Distance (EMD) between two distributions.
+  
+  Based on the POT library: https://pythonot.github.io/"""
+  import ot
+  print("Compute EMD...")
+  p, q = points[labels == 1], points[labels == -1]
+  subset_size = 1024
+  p = onp.random.permutation(p)[:subset_size]
+  q = onp.random.permutation(q)[:subset_size]
+  # uniform distribution on samples
+  M = ot.dist(p, q, metric='euclidean')
+  emd = ot.emd2([], [], M)
+  print(f"EMD (LinearSolving): {float(emd):.2f}")
+  return emd 
 
 
 def plot_data(points, labels, train_state):
@@ -178,44 +202,50 @@ def plot_data(points, labels, train_state):
 
   # make predictions of f
   Y_pred = predict_model(train_state, jnp.array(X_pred))
-  Y_pred = Y_pred.reshape(x.shape[0], y.shape[0])
+  Y_pred_square = Y_pred.reshape(x.shape[0], y.shape[0])
 
-  # plot
+  fP = predict_model(train_state, jnp.array(points[labels == 1])).flatten()
+  fQ = predict_model(train_state, jnp.array(points[labels == -1])).flatten()
+
+  # plot contours of f
   fig = plt.figure(figsize=(10, 7))
-  ax = fig.add_subplot(111)
+  ax = fig.add_subplot(121,aspect='equal')
   ax.scatter(points[labels == 1, 0], points[labels == 1, 1], alpha=0.1)
   ax.scatter(points[labels == -1, 0], points[labels == -1, 1], alpha=0.1)
-  cset = ax.contour(xx, yy, Y_pred, cmap='twilight', levels=20)
+  cset = ax.contour(xx, yy, Y_pred_square, cmap='twilight', levels=20)
   ax.clabel(cset, inline=1, fontsize=10)
+
+  # Plot histogram of fP and fQ
+  ax = fig.add_subplot(122)
+  ax.hist(fP, alpha=0.5, label='1')
+  ax.hist(fQ, alpha=0.5, label='-1')
 
   file_loc = 'sandbox/wasserstein_1.png'
   plt.savefig(file_loc)
   print(f"OT plan saved at {file_loc}.")
 
 
-
 def main(argv):
   train_ds = get_datasets()
-  rng = jax.random.PRNGKey(0)
+  rng = jax.random.PRNGKey(37)
 
   rng, init_rng = jax.random.split(rng)
   state = create_train_state(init_rng)
 
   for epoch in range(1, FLAGS.num_epochs + 1):
     rng, input_rng = jax.random.split(rng)
-    state, train_loss = train_epoch(state, train_ds,
-                                    FLAGS.batch_size,
-                                    input_rng)
-    logging.info(f'epoch:[{epoch:3d}/{FLAGS.num_epochs}], train_loss: {train_loss:.4f}')
+    state, train_loss, logs = train_epoch(state, train_ds,
+                                          FLAGS.batch_size,
+                                          input_rng)
+    logging.info(f'epoch:[{epoch:3d}/{FLAGS.num_epochs}], train_loss: {train_loss:.4f}, P: {logs[0]:.4f}, Q: {logs[1]:.4f}')
 
-  epsilon = 1e-2
-  emd = compute_earth_mover_distance(train_ds['points'], train_ds['labels'], epsilon)
   print(f"Wasserstein-1: {train_loss:.2f}")
-  print(f"Regulized EMD (ε={epsilon:.2f}): {float(emd):.2f}")
+  emd = compute_earth_mover_distance_pot(train_ds['points'], train_ds['labels'])
   error = jnp.abs(emd - train_loss) / train_loss * 100
   print(f"Relative Error: {error:.2f}%")
 
   plot_data(train_ds['points'], train_ds['labels'], state)
+
 
 if __name__ == '__main__':
   # jax.config.update("jax_enable_x64", True)
